service/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cassandra-auth-service"
version = "0.1.0"
description = "Auth service — Casbin RBAC, MCP key validation, credential storage"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0",
    "casbin>=1.36.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.15.0",
]

[project.scripts]
cassandra-auth = "cassandra_auth_service.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
cassandra_auth_service = ["../../model.conf", "../../schema.sql"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
